+++ src/components/EditSections/EditUserRoles/components/UserRolesModal/UserRolesModal.js
@@ -53,11 +53,15 @@ export default function UserRolesModal({ isOpen,
     }
   };
 
-  const toggleAllRoles = (checked) => {
+  const toggleRoleList = (checked, roleList) => {
     if (checked) {
-      setAssignedRoleIds({ ...assignedRoleIds, [tenantId]: allRolesData?.roles.map(role => role.id) });
+      const currentAssignedIds = assignedRoleIds[tenantId] || [];
+      const currentPlusFiltered = new Set([...currentAssignedIds, ...roleList.map(role => role.id)]);
+      setAssignedRoleIds({ ...assignedRoleIds, [tenantId]: Array.from(currentPlusFiltered) });
     } else {
-      setAssignedRoleIds({ ...assignedRoleIds, [tenantId]: [] });
+      const idsToUnassign = roleList.map(role => role.id);
+      const remainingIds = (assignedRoleIds[tenantId] || []).filter(id => !idsToUnassign.includes(id));
+      setAssignedRoleIds({ ...assignedRoleIds, [tenantId]: remainingIds });
     }
   };
 
@@ -172,7 +176,7 @@ export default function UserRolesModal({ isOpen,
               assignedUserRoleIds={assignedRoleIds}
               filteredRoles={filteredRoles}
               toggleRole={toggleRole}
-              toggleAllRoles={toggleAllRoles}
+              toggleRoleList={toggleRoleList}
               tenantId={tenantId}
             />
           </Pane>