+++ src/components/EditSections/EditUserRoles/components/UserRolesModal/UserRolesModal.test.js
@@ -3,15 +3,21 @@ import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import UserRolesModal from './UserRolesModal';
 
 jest.mock('../../../../../hooks', () => {
-  const mockAllRolesData = { data:{ roles: [{ id: '1', name: 'testRole' },
-    { id: '4', name: 'AAtestRole4' },
-    { id: '3', name: 'testRole3' }] },
-  allRolesMapStructure: new Map([['1', { id: '1', name: 'testRole' }],
-    ['4', { id: '4', name: 'AAtestRole4' }],
-    ['3', { id: '3', name: 'testRole3' }]]) };
+  const mockAllRolesData = {
+    data: {
+      roles: [{ id: '1', name: 'testRole' },
+        { id: '4', name: 'AAtestRole4' },
+        { id: '3', name: 'testRole3' }]
+    },
+    allRolesMapStructure: new Map([['1', { id: '1', name: 'testRole' }],
+      ['4', { id: '4', name: 'AAtestRole4' }],
+      ['3', { id: '3', name: 'testRole3' }]])
+  };
 
-  return { ...jest.requireActual('../../../../../hooks'),
-    useAllRolesData: jest.fn().mockReturnValue(mockAllRolesData) };
+  return {
+    ...jest.requireActual('../../../../../hooks'),
+    useAllRolesData: jest.fn().mockReturnValue(mockAllRolesData)
+  };
 });
 jest.unmock('@folio/stripes/components');
 
@@ -28,7 +34,7 @@ describe('UserRoleModal', () => {
     jest.clearAllMocks();
   });
   it('renders user roles modal', async () => {
-    const { getByText } = await waitFor(() => renderComponent({ isOpen: true, onClose: mockOnClose, assignedRoles: mockAssignedRoles, initialRoleIds, setAssignedRolesIds:jest.fn(), tenantId }));
+    const { getByText } = await waitFor(() => renderComponent({ isOpen: true, onClose: mockOnClose, assignedRoles: mockAssignedRoles, initialRoleIds, setAssignedRolesIds: jest.fn(), tenantId }));
 
     expect(getByText('testRole')).toBeInTheDocument();
     expect(getByText('AAtestRole4')).toBeInTheDocument();
@@ -36,18 +42,20 @@ describe('UserRoleModal', () => {
   });
 
   it('should call onClose function', async () => {
-    await waitFor(() => renderComponent({ isOpen: true, onClose: mockOnClose, assignedRoles: mockAssignedRoles, initialRoleIds, setAssignedRolesIds:jest.fn(), tenantId }));
+    await waitFor(() => renderComponent({ isOpen: true, onClose: mockOnClose, assignedRoles: mockAssignedRoles, initialRoleIds, setAssignedRolesIds: jest.fn(), tenantId }));
     await userEvent.click(document.getElementById('user-roles-modal-close-button'));
 
     expect(mockOnClose).toHaveBeenCalledTimes(1);
   });
 
   it('should select/unselect all roles, and ', async () => {
-    renderComponent({ isOpen: true,
+    renderComponent({
+      isOpen: true,
       onClose: mockOnClose,
       initialRoleIds,
-      changeUserRoles:jest.fn(),
-      tenantId });
+      changeUserRoles: jest.fn(),
+      tenantId
+    });
 
     const selectAllButton = document.querySelector('[name="selected-selectAll"]');
     await userEvent.click(selectAllButton);
@@ -62,13 +70,65 @@ describe('UserRoleModal', () => {
     expect(roleCheckbox.checked).toBe(false);
   });
 
+  it('should select all only for filtered roles and keep unfiltered assigned roles', async () => {
+    const mockChangeUserRoles = jest.fn();
+
+    renderComponent({
+      isOpen: true,
+      onClose: mockOnClose,
+      initialRoleIds: { consortium: ['1', '4'] },
+      changeUserRoles: mockChangeUserRoles,
+      tenantId,
+    });
+
+    await userEvent.type(document.querySelector('[name="query"]'), 'testRole3');
+    await userEvent.click(document.querySelector('[data-test-submit-button]'));
+
+    const selectAllButton = document.querySelector('[name="selected-selectAll"]');
+    expect(selectAllButton.checked).toBe(false);
+
+    await userEvent.click(selectAllButton);
+
+    const submitButton = document.querySelector('#clickable-permissions-modal-save');
+    await userEvent.click(submitButton);
+
+    expect(mockChangeUserRoles).toHaveBeenCalledWith(['4', '1', '3']);
+  });
+
+  it('should deselect all only for filtered roles and keep other assigned roles', async () => {
+    const mockChangeUserRoles = jest.fn();
+
+    renderComponent({
+      isOpen: true,
+      onClose: mockOnClose,
+      initialRoleIds: { consortium: ['1', '3'] },
+      changeUserRoles: mockChangeUserRoles,
+      tenantId,
+    });
+
+    await userEvent.type(document.querySelector('[name="query"]'), 'testRole3');
+    await userEvent.click(document.querySelector('[data-test-submit-button]'));
+
+    const selectAllButton = document.querySelector('[name="selected-selectAll"]');
+    expect(selectAllButton.checked).toBe(true);
+
+    await userEvent.click(selectAllButton);
+
+    const submitButton = document.querySelector('#clickable-permissions-modal-save');
+    await userEvent.click(submitButton);
+
+    expect(mockChangeUserRoles).toHaveBeenCalledWith(['1']);
+  });
+
 
   it('should show assigned roles only', async () => {
-    const { queryByText } = renderComponent({ isOpen: true,
+    const { queryByText } = renderComponent({
+      isOpen: true,
       onClose: mockOnClose,
       initialRoleIds: ['1'],
-      changeUserRoles:jest.fn(),
-      tenantId });
+      changeUserRoles: jest.fn(),
+      tenantId
+    });
 
     const assignedFilterCheckbox = document.querySelector('[name="status.assigned"]');
     await userEvent.click(assignedFilterCheckbox);
@@ -79,11 +139,13 @@ describe('UserRoleModal', () => {
   });
 
   it('should show unassigned roles only', async () => {
-    const { getAllByRole } = renderComponent({ isOpen: true,
+    const { getAllByRole } = renderComponent({
+      isOpen: true,
       onClose: mockOnClose,
       initialRoleIds,
-      changeUserRoles:jest.fn(),
-      tenantId });
+      changeUserRoles: jest.fn(),
+      tenantId
+    });
 
     const unassignedFilterCheckbox = document.querySelector('[name="status.unassigned"]');
     await userEvent.click(unassignedFilterCheckbox);
@@ -99,13 +161,15 @@ describe('UserRoleModal', () => {
   it('should reset all filters', async () => {
     const actual = jest.requireActual('./useRolesModalFilters');
     const mockFunction = jest.spyOn(actual, 'default');
-    mockFunction.mockReturnValue({ filters: {}, onChangeFilter: jest.fn(), onClearFilter: jest.fn(), resetFilters:jest.fn() });
+    mockFunction.mockReturnValue({ filters: {}, onChangeFilter: jest.fn(), onClearFilter: jest.fn(), resetFilters: jest.fn() });
 
-    renderComponent({ isOpen: true,
+    renderComponent({
+      isOpen: true,
       onClose: mockOnClose,
       initialRoleIds,
-      changeUserRoles:jest.fn(),
-      tenantId });
+      changeUserRoles: jest.fn(),
+      tenantId
+    });
 
     await userEvent.click(document.querySelector('[data-test-reset-all-button="true"]'));
 
@@ -113,11 +177,13 @@ describe('UserRoleModal', () => {
   });
 
   it('should toggle filters pane', async () => {
-    const { getByText } = renderComponent({ isOpen: true,
+    const { getByText } = renderComponent({
+      isOpen: true,
       onClose: mockOnClose,
       initialRoleIds,
-      changeUserRoles:jest.fn(),
-      tenantId });
+      changeUserRoles: jest.fn(),
+      tenantId
+    });
 
     const collapseButton = document.querySelector('[data-test-collapse-filter-pane-button="true"]');
     await userEvent.click(collapseButton);
@@ -130,11 +196,13 @@ describe('UserRoleModal', () => {
 
   it('should sort objects alphabetically, extract ids after sorting and call changeUserRoles function with that ids', async () => {
     const mockChangeUserRoles = jest.fn();
-    const { getByRole } = renderComponent({ isOpen: true,
+    const { getByRole } = renderComponent({
+      isOpen: true,
       onClose: mockOnClose,
       initialRoleIds: { 'consortium': ['1', '4'] },
-      changeUserRoles:mockChangeUserRoles,
-      tenantId });
+      changeUserRoles: mockChangeUserRoles,
+      tenantId
+    });
     const submitButton = getByRole('button', { name: 'stripes-components.saveAndClose' });
     await userEvent.click(submitButton);
     expect(mockChangeUserRoles).toHaveBeenCalledWith(['4', '1']);